implements AccessAuth on controllers

+++ src/Controller/AccessRoleController.php
@@ -3,14 +3,16 @@
 namespace App\Controller;
 
 use App\Entity\AccessRole;
+use App\Service\AccessAuth;
 use App\Form\AccessRoleType;
+use Symfony\Component\Routing\Router;
 use App\Repository\AccessRoleRepository;
+use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\Routing\Router;
 
 /**
  * @Route("/access/role")
@@ -20,8 +22,14 @@ class AccessRoleController extends AbstractController
     /**
      * @Route("/", name="access_role_index", methods={"GET"})
      */
-    public function index(AccessRoleRepository $accessRoleRepository, Request $request): Response
+    public function index(AccessRoleRepository $accessRoleRepository, AccessAuth $accessAuth, EntityManagerInterface $entityManager, Request $request)
     {
+        $redirection = $accessAuth->verif($request, $entityManager);
+        if($redirection)
+        {
+            return $this->redirect($redirection);
+        }
+
         $baseurl = $request->getScheme() . '://' . $request->getHttpHost() . $request->getBasePath() . $request->server->get('REQUEST_URI');
 
         $testpathinfo = $request->server->get('REQUEST_URI');
@@ -76,8 +84,14 @@ class AccessRoleController extends AbstractController
     /**
      * @Route("/new", name="access_role_new", methods={"GET","POST"})
      */
-    public function new(Request $request): Response
+    public function new(AccessAuth $accessAuth, EntityManagerInterface $entityManager, Request $request)
     {
+        $redirection = $accessAuth->verif($request, $entityManager);
+        if($redirection)
+        {
+            return $this->redirect($redirection);
+        }
+
         $accessRole = new AccessRole();
         $form = $this->createFormBuilder($accessRole)
             ->add('authorizationEspace', TextType::class)
@@ -105,8 +119,14 @@ class AccessRoleController extends AbstractController
     /**
      * @Route("/{id}", name="access_role_show", methods={"GET"})
      */
-    public function show(AccessRole $accessRole): Response
+    public function show(AccessRole $accessRole, AccessAuth $accessAuth, EntityManagerInterface $entityManager, Request $request)
     {
+        $redirection = $accessAuth->verif($request, $entityManager);
+        if($redirection)
+        {
+            return $this->redirect($redirection);
+        }
+
         return $this->render('access_role/show.html.twig', [
             'access_role' => $accessRole,
         ]);
@@ -115,8 +135,14 @@ class AccessRoleController extends AbstractController
     /**
      * @Route("/{id}/edit", name="access_role_edit", methods={"GET","POST"})
      */
-    public function edit(Request $request, AccessRole $accessRole): Response
+    public function edit(AccessRole $accessRole, AccessAuth $accessAuth, EntityManagerInterface $entityManager, Request $request)
     {
+        $redirection = $accessAuth->verif($request, $entityManager);
+        if($redirection)
+        {
+            return $this->redirect($redirection);
+        }
+
         $form = $this->createForm(AccessRoleType::class, $accessRole);
         $form->handleRequest($request);
 
@@ -137,8 +163,14 @@ class AccessRoleController extends AbstractController
     /**
      * @Route("/{id}", name="access_role_delete", methods={"DELETE"})
      */
-    public function delete(Request $request, AccessRole $accessRole): Response
+    public function delete(AccessRole $accessRole, AccessAuth $accessAuth, EntityManagerInterface $entityManager, Request $request)
     {
+        $redirection = $accessAuth->verif($request, $entityManager);
+        if($redirection)
+        {
+            return $this->redirect($redirection);
+        }
+
         if ($this->isCsrfTokenValid('delete'.$accessRole->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($accessRole);
